refactor test details into separate component
